+++ TBS.Data/Models/Reviews/Response/PaginatedCarrierReviews.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace TBS.Data.Models.Reviews
+{
+    public class PaginatedCarrierReviews
+    {
+        public PaginationModel PaginationModel { get; set; }
+
+        public IEnumerable<CarrierReview> CarrierReviews { get; set; }
+    }
+}