+++ TBS.Website/src/views/Delivery/Shipper/Index.vue
@@ -59,32 +59,7 @@
         </div>
       </div>
     </div>
-    <div class="row pt-3 text-center">
-      <div class="col-12 background pt-3">
-        <h4>Bid Details</h4>
-        <hr>
-        <div class="row">
-          <div class="col-12">
-            <p></p>
-            <div>
-              <table class="table">
-                <tr>
-                  <th style="width: 33.3%">Carrier</th>
-                  <th style="width: 33.3%">Bid Amount</th>
-                  <th style="width: 33.3%">Status</th>
-                </tr>
-                <tr>
-                  <td><router-link :to="{ name: 'carrierProfile', params: { id:bid.carrier.id }}" class="fade-on-hover text-blue">{{ bid.carrier.name }}</router-link></td>
-                  <td>{{ formatMoney(bid.bidAmount) }}</td>
-                  <td>{{ parseBidStatus(bid.bidStatus) }}</td>
-                </tr>
-              </table>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="row pt-3 text-center" v-if="!$store.getters['global/isLoading']">
+    <div class="row pt-3 text-center" v-if="!$store.getters['global/isLoading'] && convertedBidStatus != 'Completed'">
       <div class="col-12 pt-3 pb-3">
         <div class="fixed-bottom pb-btn text-center">
           <div v-if="accountType == 'carrier'">
@@ -94,6 +69,62 @@
           <div v-else>
             <button v-if="convertedBidStatus == 'Pending Delivery'" type="button" class="btn btn-main bg-blue fade-on-hover text-uppercase text-white" @click="updateBid('Completed')">Force Delivery</button>
             <button v-if="convertedBidStatus == 'Pending Delivery Approval'" type="button" class="btn btn-main bg-blue fade-on-hover text-uppercase text-white" @click="updateBid('Completed')">Approve Delivery</button>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="row pt-3 text-center" v-if="!$store.getters['global/isLoading'] && convertedBidStatus != 'Completed'">
+      <div class="col-12 pt-3 pb-3">
+        <div class="fixed-bottom pb-btn text-center">
+          <div v-if="accountType == 'carrier'">
+            <button v-if="convertedBidStatus == 'Pending Delivery'" type="button" class="btn btn-main bg-blue fade-on-hover text-uppercase text-white" @click="updateBid('Pending Delivery Approval')">Confirm Delivery</button>
+            <button v-if="convertedBidStatus == 'Pending Delivery Approval'" type="button" class="btn btn-main bg-blue fade-on-hover text-uppercase text-white" @click="sendReminder()">Send Reminder</button>
+          </div>
+          <div v-else>
+            <button v-if="convertedBidStatus == 'Pending Delivery'" type="button" class="btn btn-main bg-blue fade-on-hover text-uppercase text-white" @click="updateBid('Completed')">Force Delivery</button>
+            <button v-if="convertedBidStatus == 'Pending Delivery Approval'" type="button" class="btn btn-main bg-blue fade-on-hover text-uppercase text-white" @click="updateBid('Completed')">Approve Delivery</button>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="row pt-3 text-center" v-if="convertedBidStatus === 'Completed'">
+      <div class="col-12 background">
+        <p></p>
+        <h4>Review</h4>
+        <hr>
+        <div class="row pt-3" v-if="carrierReview == null && accountType == 'carrier' && !carrierReviewPlaced">
+          <div class="col-12 pt-3 pb-3"> 
+            <h5>Place a Review</h5>
+            <p><star-rating style="display: inline-block;" :increment="0.5" @rating-selected="setCarrierReview" :show-rating=false :star-size=30></star-rating></p>
+            <span>Enter comment here </span>
+            <br>
+            <textarea style="min-width:35%" v-model="tempCarrierReview.review"/><br><br>
+            <button type="button" class="btn btn-main bg-blue fade-on-hover text-uppercase text-white mb-3" @click="createReview('carrier', tempCarrierReview)">Enter Review</button><br>
+          </div>
+        </div>
+        <div class="row pt-3" v-if="shipperReview == null && accountType == 'shipper' && !shipperReviewPlaced">
+          <div class="col-12 pt-3 pb-3"> 
+            <h5>Place a Review</h5>
+            <p><star-rating style="display: inline-block;" :increment="0.5" @rating-selected="setShipperReview" :show-rating=false :star-size=30></star-rating></p>
+            <span>Enter comment here </span>
+            <br>
+            <textarea style="min-width:35%" v-model="tempShipperReview.review"/><br><br>
+            <button type="button" class="btn btn-main bg-blue fade-on-hover text-uppercase text-white mb-3" @click="createReview('shipper', tempShipperReview)">Enter Review</button><br>
+          </div>
+        </div>
+        <div class="row pt-3">
+          <div class="col-12" v-if="carrierReview != null">
+            <p class="font-weight-bold">Review from Carrier<p>
+            <p>{{ carrierReview.shipper.name }} | {{`${parseDate(carrierReview.reviewDate)}` }}</p>
+            <p><star-rating style="display: inline-block;" :increment="0.5" :show-rating=false :read-only=true :star-size=30 v-model="carrierReview.rating"></star-rating></p>
+            <p>{{ carrierReview.review }}</p>
+          </div>
+          <hr v-if="carrierReview && shipperReview" class="pb-3">
+          <div class="col-12" v-if="shipperReview != null">
+            <p class="font-weight-bold">Review from Shipper<p>
+            <p>{{ shipperReview.carrier.name }} |  {{`${parseDate(shipperReview.reviewDate)}` }}</p>
+            <p><star-rating style="display: inline-block;" :increment="0.5" :show-rating=false :read-only=true :star-size=30 v-model="shipperReview.rating"></star-rating></p>
+            <p>{{ shipperReview.review }}</p>
           </div>
         </div>
       </div>
@@ -108,11 +139,12 @@ import Swal from 'sweetalert2'
 
 import postUtilities from '@/utils/postUtilities.js'
 import bidUtilities from '@/utils/bidUtilities.js'
-
+import StarRating from 'vue-star-rating'
 export default {
   name: 'shipperDelivery',
   components: {
-    Back
+    Back,
+    StarRating
   },
   props: {
     postId: String,
@@ -120,10 +152,23 @@ export default {
   },
   data() {
     return {
+      bidType: false,
       accountType: '',
       postType: '',
       post: null,
       bid: null,
+      carrierReviewPlaced: false,
+      shipperReviewPlaced: false,
+      carrierReview: null,
+      tempCarrierReview: {
+        rating: 1,
+        review: ''
+      },
+      shipperReview: null,
+      tempShipperReview: {
+        rating: 1,
+        review: ''
+      },
       error: null,
       convertedBidStatus: null,
     }
@@ -135,6 +180,8 @@ export default {
           this.post = response.data.result
           this.bid = this.post.bids.find((bid) => bid.id == this.bidId)
           this.convertedBidStatus = this.parseBidStatus(this.bid.bidStatus)
+          this.carrierReview = this.bid.carrierReview
+          this.shipperReview = this.bid.shipperReview
         })
         .catch(() => {
           Swal.fire({
@@ -180,6 +227,50 @@ export default {
             text: 'Something went wrong! Please try again!',
           })
         })
+    },     
+    createReview(type, tempReview) {
+      let review = {
+        bidId: this.bid.id,
+        bidBoolean: this.bidType,
+        rating: tempReview.rating,
+        review: tempReview.review
+      }
+      this.$store
+        .dispatch('bids/createReview', {
+          type: type,
+          Review: review
+        })
+        .then(() => {
+          if (type == 'carrier') {
+            this.carrierReview = review
+            this.carrierReview.shipper = this.post.shipper
+            this.carrierReview.reviewDate = new Date()
+            this.carrierReviewPlaced = true
+          } else {
+            this.shipperReview = review
+            this.shipperReview.carrier = this.bid.carrier
+            this.shipperReview.reviewDate = new Date()
+            this.shipperReviewPlaced = true
+          }
+          Swal.fire({
+            type: 'success',
+            title: 'Completed',
+            text: 'The review has been created!'
+          })
+        })
+        .catch(() => {
+          Swal.fire({
+            type: 'error',
+            title: 'Oops...',
+            text: 'Something went wrong! Please try again!'
+          })
+        })
+    },
+    setCarrierReview(rating) {
+      this.tempCarrierReview.rating = rating
+    },
+    setShipperReview(rating) {
+      this.tempShipperReview.rating = rating
     },
     parsePostStatus(status) {
       return postUtilities.parsePostStatus(status)