+++ TBS.Website/src/views/Posts/Carrier/Details.vue
@@ -50,7 +50,7 @@
                 </tr>
                 <tr>
                   <td><router-link :to="{ name: 'carrierProfile', params: {id:post.carrier.id}}" class="fade-on-hover text-blue">{{ post.carrier.name }}</router-link></td>
-                  <td>Coming Soon..</td>
+                  <td><star-rating :inline=true :show-rating=false :increment="0.5" :read-only=true :star-size=30 v-model="reviewScore"></star-rating></td>
                   <td><a :href="'mailto:' + post.carrier.email">{{ post.carrier.email }}</a></td>
                 </tr>
               </table>
@@ -69,7 +69,7 @@
                       <th style="width: 25%">Trailer Type</th>
                     </tr>
                     <tr>
-                      <td>Coming soon</td> 
+                      <td>{{ `${post.carrier.vehicle.year} ${post.carrier.vehicle.make} ${post.carrier.vehicle.model}` }}</td> 
                       <td>{{ post.spacesAvailable }}</td>
                       <td>{{ parseTrailerType(post.trailerType) }}</td>
                     </tr>
@@ -89,7 +89,7 @@
               <table class="table">
                 <tr>
                   <th style="width: 33.3%">Date Posted</th>
-                  <th style="width: 33.3%">Starting Bid</th>
+                  <th style="width: 33.3%">Minimum Bid</th>
                   <th style="width: 33.3%">Highest Bid</th>
                 </tr>
                 <tr>
@@ -204,7 +204,7 @@ import TextInput from '@/components/Form/Input/TextInput.vue'
 import DateInput from '@/components/Form/Input/DateInput.vue'
 import ConditionInput from '@/components/Form/Input/ConditionInput.vue'
 import YearInput from '@/components/Form/Input/YearInput.vue'
-
+import StarRating from "vue-star-rating";
 import postUtilities from '@/utils/postUtilities.js'
 import { required, helpers } from 'vuelidate/lib/validators'
 const bidRegex = helpers.regex('bidRegex', /^[+]?([0-9]+(?:[.][0-9]*)?|\.[0-9]+)$/)
@@ -216,6 +216,7 @@ export default {
     DateInput,
     ConditionInput,
     YearInput,
+    StarRating,
   },
   props: {
     id: String
@@ -256,6 +257,7 @@ export default {
       toShowModal: false,
       validPickupAddress: null,
       validDropoffAddress: null,
+      reviewScore: 0,
     }
   },
   methods: {
@@ -273,6 +275,16 @@ export default {
               max = (v > max) ? v : max;
             }
             this.highestBid = max
+          }
+          let reviews = this.post.carrier.reviews
+          this.reviewScore = 0
+          if (reviews !== null) {            
+            var totalReviews = 0
+            reviews.forEach(review => {
+              this.reviewScore += review.rating
+              totalReviews += 1
+            })
+            this.reviewScore = this.reviewScore / totalReviews
           }
         })
         .catch(() => {
@@ -337,11 +349,12 @@ export default {
     },
     parsePickupLocationAddress(address) {
       try {
-        this.bidPost.pickupLocation.addressLine = `${address[0].long_name} ${address[1].long_name}`
-        this.bidPost.pickupLocation.city = address[2].long_name
-        this.bidPost.pickupLocation.province = address[4].short_name
-        this.bidPost.pickupLocation.country = address[5].long_name
-        this.bidPost.pickupLocation.postalCode = address[6].long_name
+        // TODO FIX
+        this.bidPost.pickupLocation.addressLine = `${address.find(a => a.types.includes("street_number")).short_name} ${address.find(a => a.types.includes("route")).long_name}`
+        this.bidPost.pickupLocation.city = address.find(a => a.types.includes("locality")).long_name
+        this.bidPost.pickupLocation.province = address.find(a => a.types.includes("administrative_area_level_1")).short_name
+        this.bidPost.pickupLocation.country = address.find(a => a.types.includes("country")).long_name
+        this.bidPost.pickupLocation.postalCode = address.find(a => a.types.includes("postal_code")).long_name
         this.validPickupAddress = true
       } catch {
         this.validPickupAddress = false