+++ TBS.Website/src/views/Profile/Shipper.vue
@@ -7,67 +7,71 @@
         <h6 @click="$router.go(-1)" class="btn btn-main bg-blue fade-on-hover text-uppercase text-white">Click here to return</h6>
       </div>
     </div>
-    <div class="row pt-3 text-center">
-      <div class="col-12">
-        <h3>{{profile.name}}</h3>
+    <div v-if="profile">
+      <div class="row pt-3 text-center">
+        <div class="col-12">
+          <h3>{{ profile.name }}</h3>
+        </div>
       </div>
-    </div>
-    <div class="row pt-3 text-center">
-      <div class="col-12 background">
-        <div class="row pt-3">
-          <div class="col-12">
-            <table class="table">
-              <tr>
-                <th style="width: 50%">Email</th>
-                <th style="width: 50%">Dealer Number</th>
-              </tr>
-              <tr>
-                <td><a :href="'mailto:' + profile.email">{{ profile.email }}</a></td> 
-                <td>{{ profile.dealerNumber }}</td>
-              </tr>
-            </table>
+      <div class="row pt-3 text-center">
+        <div class="col-12 background">
+          <div class="row pt-3">
+            <div class="col-12">
+              <table class="table">
+                <tr>
+                  <th style="width: 50%">Email</th>
+                  <th style="width: 50%">Dealer Number</th>
+                </tr>
+                <tr>
+                  <td><a :href="'mailto:' + profile.email">{{ profile.email }}</a></td> 
+                  <td>{{ profile.dealerNumber }}</td>
+                </tr>
+              </table>
+            </div>
           </div>
         </div>
       </div>
-    </div>
-    <div class="row pt-3 text-center">
-      <div class="col-12">
-        <h3>{{ profile.company.name }}</h3>
-        <h6>{{ formatAddress(profile.company.address) }}</h6>
+      <div class="row pt-3 text-center">
+        <div class="col-12">
+          <h3>{{ profile.company.name }}</h3>
+          <h6>{{ formatAddress(profile.company.address) }}</h6>
+        </div>
       </div>
-    </div>
-    <div class="row pt-3 text-center">
-      <div class="col-12 background">
-        <div class="row pt-3">
-          <div class="col-12">
-            <table class="table">
-              <tr>
-                <th style="width: 33.3%"></th>
-                <th style="width: 33.3%">Company Email</th>
-                <th style="width: 33.3%">Company Phone</th>
-              </tr>
-              <tr>
-                <td>{{ profile.company.contact.name }}</td> 
-                <td><a :href="'mailto:' + profile.company.contact.email">{{ profile.company.contact.email }}</a></td> 
-                <td><a :href="'tel:' + profile.company.contact.phoneNumber">{{ profile.company.contact.phoneNumber }}</a></td> 
-              </tr>
-            </table>
+      <div class="row pt-3 text-center">
+        <div class="col-12 background">
+          <div class="row pt-3">
+            <div class="col-12">
+              <table class="table">
+                <tr>
+                  <th style="width: 33.3%"></th>
+                  <th style="width: 33.3%">Company Email</th>
+                  <th style="width: 33.3%">Company Phone</th>
+                </tr>
+                <tr>
+                  <td>{{ profile.company.contact.name }}</td> 
+                  <td><a :href="'mailto:' + profile.company.contact.email">{{ profile.company.contact.email }}</a></td> 
+                  <td><a :href="'tel:' + profile.company.contact.phoneNumber">{{ profile.company.contact.phoneNumber }}</a></td> 
+                </tr>
+              </table>
+            </div>
           </div>
         </div>
       </div>
-    </div>
-    <div class="row pt-3 text-center">
-      <div class="col-12">
-        <h3>Reviews</h3>
-      </div>     
-    </div>
-    <div class="row pt-3">
-      <div class="col-12 background">
-        <div class="row pt-3">
-          <div class="col-12" v-for="(review, index) in reviews" :key="index">
-            <p>{{ review.rating }}</p>
-            <p>{{ review.carrier.name }} | {{ review.date }}</p>
-            <p>{{ review.comment }}</p>
+      <div class="row pt-3 text-center">
+        <div class="col-12">
+          <h3 v-if="reviews.length != 0">Reviews</h3>
+          <h3 v-else>No Reviews</h3>
+        </div>     
+      </div>
+      <div class="row pt-3" v-if="reviews.length != 0">
+        <div class="col-12 background">
+          <div class="row pt-3">
+            <div class="col-12 text-center" v-for="(review, index) in reviews" :key="index">
+              <p>{{ review.carrier.name }} | {{`${parseDate(review.reviewDate)}` }}</p>
+              <p><star-rating style="display: inline-block;" :increment="0.5" :show-rating=false :read-only=true :star-size=30 v-model="review.rating"></star-rating></p>
+              <p>{{ review.review }}</p>
+              <hr v-if="reviews.length - 1 != index" class="pt-3">
+            </div>
           </div>
         </div>
       </div>
@@ -79,41 +83,28 @@
 import Swal from 'sweetalert2'
 
 import postUtilities from '@/utils/postUtilities.js'
-
+import StarRating from 'vue-star-rating'
 export default {
   name: 'detailedCarrierPost',
   components: {
+    StarRating
   },
   props: {
     id: String
   },
   data() {
     return {
-      profile: {
-        vehicle: {
-          year: new Date().getUTCFullYear().toString(),
-          make: '',
-          model: '',
-          VIN: '',
-          condition: 'New',
-        },       
-      },
-      reviews:[
-          {
-            rating: '3.5',
-            comment: "This is a test comment",
-            date: "10-20-2018",
-            carrier: {
-              name: "Testing person"
-            },
-          }
-        ],
+      profile: null,
+      reviews: [],
       error: false,
     }
   },
   methods: {
     parseTrailerType(trailerType) {
       return postUtilities.parseTrailerType(trailerType)
+    },
+     parseDate(value) {
+      return postUtilities.parseDate(value)
     },
     formatAddress(address) {
       return postUtilities.formatAddress(address)
@@ -122,6 +113,7 @@ export default {
       this.$store.dispatch('profiles/getProfileById', {profileId: this.id, type: 'shipper'})
         .then((response) => {
           this.profile = response.data.result
+          this.reviews = this.profile.reviews
         })
         .catch(() => {
           Swal.fire({